Refactor marquee setup for multi-line TextViews
