+++ app/src/main/java/org/xbmc/kore/ui/AbstractInfoFragment.java
@@ -35,7 +35,6 @@ import android.support.v4.app.FragmentManager;
 import android.support.v4.content.ContextCompat;
 import android.support.v4.widget.NestedScrollView;
 import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v4.widget.TextViewCompat;
 import android.text.TextUtils;
 import android.util.DisplayMetrics;
 import android.view.LayoutInflater;
@@ -375,31 +374,7 @@ abstract public class AbstractInfoFragment extends AbstractFragment
     @SuppressLint("StringFormatInvalid")
     protected void updateView(DataHolder dataHolder) {
         titleTextView.setText(dataHolder.getTitle());
-        titleTextView.post(new Runnable() {
-            @Override
-            public void run() {
-                int lines = titleTextView.getLineCount();
-                int maxLines = TextViewCompat.getMaxLines(titleTextView);
-                if (lines > maxLines) {
-                    titleTextView.setEllipsize(TextUtils.TruncateAt.END);
-                    titleTextView.setClickable(true);
-                    titleTextView.setOnClickListener(new View.OnClickListener() {
-                        @Override
-                        public void onClick(View v) {
-                            v.setSelected(!v.isSelected());
-                            TextUtils.TruncateAt ellipsize;
-                            if (v.isSelected()) {
-                                ellipsize = TextUtils.TruncateAt.MARQUEE;
-                            } else {
-                                ellipsize = TextUtils.TruncateAt.END;
-                            }
-                            titleTextView.setEllipsize(ellipsize);
-                            titleTextView.setHorizontallyScrolling(v.isSelected());
-                        }
-                    });
-                }
-            }
-        });
+        titleTextView.post(UIUtils.getMarqueeToggleableAction(titleTextView));
         underTitleTextView.setText(dataHolder.getUnderTitle());
         detailsRightTextView.setText(dataHolder.getDetails());
